Commit homework/hw08 homework/hw09 in-class/0215

+++ in-class/0215/activity12a.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 using namespace std;
 
-int getSmaller(int num1, int num2) {
-  int smaller;
+double getSmaller(double num1, double num2) {
+  double smaller;
   if (num1 < num2) {
     smaller = num1;
   } else {
@@ -12,9 +12,9 @@ int getSmaller(int num1, int num2) {
 }
 
 int main() {
-  int userNum1, userNum2;
+  double userNum1, userNum2;
   cout << "Enter two numbers separated by a space: ";
   cin >> userNum1 >> userNum2;
-  int smallerNum = getSmaller(userNum1, userNum2);
+  double smallerNum = getSmaller(userNum1, userNum2);
   cout << "The smaller number is " << smallerNum << endl;
 }
